myPage 분리하기
